+++ src/util.cc
@@ -106,30 +106,28 @@ void Split(std::string in, std::string delim, std::vector<std::string> *out) {
   } while(pos != std::string::npos);
 }
 
-void BytesToHuman(char *s, unsigned long long n) {
+void BytesToHuman(char *s, uint64_t n) {
   double d;
 
   if (n < 1024) {
-    /* Bytes */
     sprintf(s,"%lluB",n);
     return;
   } else if (n < (1024*1024)) {
-    d = (double)n/(1024);
+    d = static_cast<double>(n)/(1024);
     sprintf(s,"%.2fK",d);
   } else if (n < (1024LL*1024*1024)) {
-    d = (double)n/(1024*1024);
+    d = static_cast<double>(n)/(1024*1024);
     sprintf(s,"%.2fM",d);
   } else if (n < (1024LL*1024*1024*1024)) {
-    d = (double)n/(1024LL*1024*1024);
+    d = static_cast<double>(n)/(1024LL*1024*1024);
     sprintf(s,"%.2fG",d);
   } else if (n < (1024LL*1024*1024*1024*1024)) {
-    d = (double)n/(1024LL*1024*1024*1024);
+    d = static_cast<double>(n)/(1024LL*1024*1024*1024);
     sprintf(s,"%.2fT",d);
   } else if (n < (1024LL*1024*1024*1024*1024*1024)) {
-    d = (double)n/(1024LL*1024*1024*1024*1024);
+    d = static_cast<double>(n)/(1024LL*1024*1024*1024*1024);
     sprintf(s,"%.2fP",d);
   } else {
-    /* Let's hope we never need this */
     sprintf(s,"%lluB",n);
   }
 }
